Consolidate misplaced tests into tests/unit/test_config.py
